Move files into sub packages
